# components/swaggerDocumentation.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "Job management API collection",
        "version": "1.0.0",
        "title": "Job management API"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Local Url"
        },
        {
            "url": "https://28pefee5bg.execute-api.ap-south-1.amazonaws.com/dev/api/v1",
            "description": "Development Url"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
...
    "paths": {
...
        "/projects/{id}": {
            "post": {
                "summary": "To facilitate posting client's requirement.",
                "description": "To facilitate posting client's requirement. id passed as base64encoded.",
                "tags": [
                    "Job"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "Provide job Id as base64encoded",
                        "example": "NjMyMDdmZjQyMjFlMmYxZWFmZmJiZTIw"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/JobPosting"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/successResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/components/schemas/invalidResponse"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "To facilitate updating client's requirement.",
                "description": "To facilitate updating client's requirement. id passed as base64encoded.",
                "tags": [
                    "Job"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "Provide job Id as base64encoded",
                        "example": "NjM2Yjk3NjgzNDBmMmU1ZWUyYTBiM2Yw"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/JobPosting"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/successResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/components/schemas/invalidResponse"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get": {
                "summary": "To facilitate fetching client's requirement.",
                "description": "To facilitate fetching client's requirement. id passed as base64encoded.",
                "tags": [
                    "Job"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "Provide job Id as base64encoded",
                        "example": "NjM2Yjk3NjgzNDBmMmU1ZWUyYTBiM2Yw"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/successResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/components/schemas/invalidResponse"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "successResponse": {
                "type": "object",
                "properties": {
                    "successCode": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "invalidResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "JobPosting": {
                "type": "object",
                "properties": {
                    "client_name": {
                        "type": "string",
                        "example": "Demo company"
                    },
                    "client_poc": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/POC"
                        }
                    },
                    "role": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "experience_range": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExpRange"
                        }
                    },
                    "client_price": {
                        "type": "integer",
                        "example": 32
                    },
                    "ss_price": {
                        "type": "integer",
                        "example": 35
                    },
                    "month_of_engagement": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EngageMonth"
                        }
                    },
                    "engagement_type": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EngageType"
                        }
                    },
                    "no_requirements": {
                        "type": "integer",
                        "example": 2
                    },
                    "tentative_start": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tentative"
                        }
                    },
                    "expectations": {
                        "type": "string",
                        "example": "Automation Testing"
                    },
                    "primary_skills": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PSkills"
                        }
                    },
                    "secondary_skills": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SSkills"
                        }
                    },
                    "working_time_zone": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TZ"
                        }
                    },
                    "working_hours": {
                        "type": "string",
                        "example": "2"
                    },
                    "travel_preference": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TravelPref"
                        }
                    },
                    "tools_used": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tools"
                        }
                    },
                    "system_provided": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Sys"
                        }
                    },
                    "no_of_rounds": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_responsibility": {
                        "type": "string",
                        "example": "<p>Automation Testing!!</p>"
                    },
                    "is_world_wide": {
                        "type": "boolean",
                        "example": false
                    },
                    "location": {
                        "type": "string",
                        "example": "Indore, WV, USA"
                    },
                    "job_city": {
                        "type": "string",
                        "example": "Indore"
                    },
                    "job_state": {
                        "type": "string",
                        "example": "MP"
                    },
                    "job_country": {
                        "type": "string",
                        "example": "India"
                    },
                    "interview_rounds": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Interview"
                        }
                    },
                    "communication_skill": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommSkills"
                        }
                    }
                }
            },
...
            "TravelPref": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631066f661108404f9b19068"
                    },
                    "data": {
                        "type": "string",
                        "example": "Onsite"
                    }
                }
            },
            "POC": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "6322933642eb1f964d56a6ed"
                    },
                    "email": {
                        "type": "string",
                        "example": "sachins@gmail.coms"
                    },
                    "mobile_number": {
                        "type": "string",
                        "example": "+18889000183"
                    },
                    "client_poc": {
                        "type": "string",
                        "example": "Sachin Mane"
                    },
                    "designation": {
                        "type": "string",
                        "example": "Team Lead"
                    },
                    "comment": {
                        "type": "string",
                        "example": "Doc for Job Management"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631ec339f354aa036d5c1ac4"
                    },
                    "role": {
                        "type": "string",
                        "example": "Backend Engineer"
                    },
                    "skillsArray": {
                        "type": "array",
                        "item": {
                            "$ref": "#/components/schemas/SkillsArr"
                        }
                    }
                }
            },
            "SkillsArr": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631ee0f7f354aa036d5c1b00"
                    },
                    "skill": {
                        "type": "string",
                        "example": "PHP"
                    }
                }
            },
            "EngageType": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "6310439361108404f9b18fcb"
                    },
                    "data": {
                        "type": "string",
                        "example": "6 Months"
                    }
                }
            },
            "ExpRange": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "630dd9b961108404f9b18f6c"
                    },
                    "data": {
                        "type": "string",
                        "example": "7+"
                    }
                }
            },
            "EngageMonth": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "6310468b61108404f9b18fcc"
                    },
                    "data": {
                        "type": "string",
                        "example": "12 Months"
                    }
                }
            },
            "Tentative": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "6310602b61108404f9b18fe9"
                    },
                    "data": {
                        "type": "string",
                        "example": "Not Specified"
                    }
                }
            },
            "PSkills": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631ee0f7f354aa036d5c1b01"
                    },
                    "skill": {
                        "type": "string",
                        "example": "Typescript"
                    },
                    "competency": {
                        "type": "array",
                        "item": {
                            "$ref": "#/components/schemas/Competency"
                        }
                    }
                }
            },
            "Competency": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "6310606561108404f9b18fed"
                    },
                    "data": {
                        "type": "string",
                        "example": "Advanced"
                    }
                }
            },
            "SSkills": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631ee0f7f354aa036d5c1b04"
                    },
                    "skill": {
                        "type": "string",
                        "example": "Typescript"
                    }
                }
            },
            "TZ": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "632417a9b1d7dabe7f5730f6"
                    },
                    "label": {
                        "type": "string",
                        "example": "Africa/Abidjan( UTC -00:16 )"
                    },
                    "data": {
                        "type": "string",
                        "example": "Africa/Abidjan"
                    }
                }
            },
            "Interview": {
                "type": "object",
                "properties": {
                    "round_number": {
                        "type": "integer",
                        "example": 1
                    },
                    "round_name": {
                        "type": "string",
                        "example": "Technical Round"
                    }
                }
            },
            "Tools": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631071ef61108404f9b1907b"
                    },
                    "data": {
                        "type": "string",
                        "example": "Agile Practice"
                    }
                }
            },
            "Sys": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "631072ea61108404f9b19085"
                    },
                    "data": {
                        "type": "string",
                        "example": "Yes"
                    }
                }
            },
            "CommSkills": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "63294c345165559551a5e64d"
                    },
                    "data": {
                        "type": "string",
                        "example": "Average"
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}